refactor(model): make abicheck.model a package and split fact dataclasses from their parsers
ADR-061 Phase 5. `abicheck/model.py` becomes `abicheck/model/`, split by
responsibility into `vocabulary`, `declarations`, `entities`,
`extraction_contract`, `snapshot` and `stdlib_surface` — none above the
800-line ceiling, and the flat module's debt entry is retired rather than
lowered. Every name the flat module exported still resolves from
`abicheck.model`, pinned by `__all__` and by
`tests/test_model_package_surface.py`.

Creating the package forced the split Phase 4 was blocked on rather than
merely permitting it: `abicheck/model.py` and `abicheck/model/` cannot
coexist, and the flat module imported `elf_metadata` at runtime for
`SymbolBinding`. So eleven modules that conflated a model dataclass with its
parser — elf/pe/macho/dwarf/dwarf_advanced/sycl/symvers metadata,
python_api, python_ext, numpy_capi, build_mode — keep their parsing and hand
their dataclasses to `model/*_facts.py`. Each parser imports and re-exports
its own types, so `from abicheck.elf_metadata import ElfMetadata` and every
sibling spelling are unchanged; internal `abicheck/` callers now import the
model owner directly.

`AbiSnapshot.index()` carried three copies of "index by key, first wins, warn
about what was dropped". That is now one `model/first_wins_index.py`
primitive with its contract stated as Hypothesis invariants, per AGENTS.md's
rule for a reusable dedupe helper. Behaviour is unchanged, warning wording
and counts included.

One residual is recorded rather than papered over: `AbiSnapshot.build_source`
is typed `BuildSourcePack`, whose persistence half belongs to `storage`,
which `model` may not import. `buildsource/pack.py` is classified `model` by
target owner with the split noted as still owed in the ADR.
