+++ abicheck/ctf_metadata.py
@@ -41,7 +41,7 @@ import zlib
 from dataclasses import dataclass, field
 from pathlib import Path
 
-from .dwarf_metadata import DwarfMetadata, EnumInfo, FieldInfo, StructLayout
+from .model.dwarf_facts import DwarfMetadata, EnumInfo, FieldInfo, StructLayout
 from .type_metadata import FuncProto, read_null_terminated_string
 
 log = logging.getLogger(__name__)
