+++ abicheck/diff_wheel_deployment.py
@@ -44,8 +44,8 @@ from .checker_policy import ChangeKind
 from .checker_types import Change
 from .diff_helpers import make_change
 from .diff_versioning import _parse_dotted_numeric_version, _version_le
-from .elf_metadata import ElfMetadata
-from .macho_metadata import MachoMetadata
+from .model.elf_facts import ElfMetadata
+from .model.macho_facts import MachoMetadata
 
 #: Declared-floor key for :func:`check_macos_deployment_target_floor`, in the
 #: same ``runtime_floors``/``EnvironmentMatrix`` mapping G10/G27's other
