+++ abicheck/model/snapshot.py
@@ -1,4 +1,5 @@
 # Copyright 2026 Nikolay Petrov
+# SPDX-License-Identifier: Apache-2.0
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,508 +13,42 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""ABI data model — shared across dumper, checker and reporter."""
+"""``AbiSnapshot`` — the complete captured ABI surface of one library build.
+
+Every extraction layer (L0 symbols through L5 source graph) writes its facts
+into one of this dataclass's fields, and every comparison, policy and report
+stage reads them back out. It owns the shape of a snapshot only: producing one
+belongs to ``extract``, persisting one to ``storage``.
+"""
 
 from __future__ import annotations
 
 import logging as _logging
 from dataclasses import dataclass, field
-from enum import Enum
 from typing import TYPE_CHECKING
 
-# Symbol *linkage* (GLOBAL/WEAK/LOCAL/UNIQUE/OTHER, from ELF st_info.bind) —
-# reused directly from elf_metadata rather than duplicated as a second
-# model-local enum: elf_metadata.py has no local imports of its own (verified:
-# it imports only stdlib + pyelftools), so importing it here carries no cycle
-# risk, unlike ElfVisibility/Visibility above which predate this and stayed
-# model-local. binder.py separately defines its own, unrelated
-# ``SymbolBinding`` dataclass (a consumer-side *import resolution* record, not
-# a linkage classification) — that module already resolves the name collision
-# via ``as ElfSymbolBinding``; nothing here needs the same aliasing since this
-# module never imports binder.py.
-from .elf_metadata import SymbolBinding as SymbolBinding
-
-# Re-export the name-classification predicates (moved to name_classification in
-# C10) under their historical names. Redundant ``as`` aliases are the explicit
-# re-export idiom mypy recognises, so ``from .model import is_non_abi_surface_type``
-# keeps type-checking cleanly for the ~9 detector modules that use it.
-from .name_classification import (
-    COMPILER_INTERNAL_TYPES as COMPILER_INTERNAL_TYPES,
-    canonicalize_type_name as canonicalize_type_name,
-    cv_qualifiers_only_differ as cv_qualifiers_only_differ,
-    func_signature_cv_only_differ as func_signature_cv_only_differ,
-    is_abi_surface_type_name as is_abi_surface_type_name,
-    is_compiler_internal_type as is_compiler_internal_type,
-    is_cxx_runtime_library as is_cxx_runtime_library,
-    is_non_abi_surface_type as is_non_abi_surface_type,
-)
+from .declarations import Function, Variable
+from .entities import EnumType, RecordType
+from .extraction_contract import DependencyInfo, ExtractionContract
+from .first_wins_index import build_first_wins_index, describe_dropped
 
 if TYPE_CHECKING:
-    from .build_mode import BuildMode
-    from .buildsource.model import BuildSourceRef
-    from .buildsource.pack import BuildSourcePack
-    from .dwarf_advanced import AdvancedDwarfMetadata
-    from .dwarf_metadata import DwarfMetadata
-    from .elf_metadata import ElfMetadata
-    from .macho_metadata import MachoMetadata
-    from .numpy_capi import NumPyCapiSurface
-    from .pe_metadata import PeMetadata
-    from .python_api import PythonApiSurface
-    from .python_ext import PythonExtMetadata
-    from .sycl_metadata import SyclMetadata
-    from .symvers_metadata import KabiMetadata
+    from ..buildsource.model import BuildSourceRef
+    from ..buildsource.pack import BuildSourcePack
+    from .build_mode_facts import BuildMode
+    from .dwarf_facts import AdvancedDwarfMetadata, DwarfMetadata
+    from .elf_facts import ElfMetadata
+    from .kabi_facts import KabiMetadata
+    from .macho_facts import MachoMetadata
+    from .pe_facts import PeMetadata
+    from .python_facts import (
+        NumPyCapiSurface,
+        PythonApiSurface,
+        PythonExtMetadata,
+    )
+    from .sycl_facts import SyclMetadata
 
 _model_log = _logging.getLogger(__name__)
-
-# ---------------------------------------------------------------------------
-# Name classification (FIX-D) — single source of truth.
-# The pure name → bool predicates now live in name_classification (C10) so the
-# symbol-name and type-name classifiers share one home; they are re-exported
-# from this module (see the imports at the top) under their historical names
-# because ~9 detector modules import them ``from .model``. The snapshot-aware
-# wrapper below (stdlib_namespaces_excluded) stays in model.
-# ---------------------------------------------------------------------------
-
-
-def stdlib_namespaces_excluded(old: AbiSnapshot, new: AbiSnapshot) -> bool:
-    """Return True when ``std::``/runtime namespaces should be filtered out of
-    type diffing as leaked dependencies.
-
-    False only when *either* side IS the C++ runtime (libstdc++ / libc++), where
-    those types are the surface under test.  Single source of truth so every
-    registered detector that consumes ``snapshot.types`` agrees on whether to
-    keep std:: records (validation/REPORT.md FP-1; Codex reviews on PR #273).
-
-    Note (cross-implementation comparisons): when two snapshots are built
-    against *different* stdlib implementations (libstdc++ ↔ libc++), standalone
-    ``std::`` records in debug info differ wholesale even when the public ABI
-    does not embed them — so this filter stays ON to avoid flooding BREAKING
-    findings for toolchain-owned internals. The cross-implementation hazard is
-    surfaced instead by the build-mode diff (``diff_stdlib_impl.py``) as a RISK
-    finding, and a public owner type that *does* embed a ``std::`` type by value
-    is caught through its own (non-``std::``, never-filtered) layout change.
-    Per-owner un-filtering of the specific embedded records is deferred to the
-    layout-closure work.
-    """
-    old_elf = getattr(old, "elf", None)
-    new_elf = getattr(new, "elf", None)
-    return not (
-        is_cxx_runtime_library(old.library)
-        or is_cxx_runtime_library(new.library)
-        or is_cxx_runtime_library(getattr(old_elf, "soname", ""))
-        or is_cxx_runtime_library(getattr(new_elf, "soname", ""))
-    )
-
-
-# Type-name canonicalization and cv-qualifier helpers
-# (``canonicalize_type_name`` / ``cv_qualifiers_only_differ``) now live in the
-# dependency-free ``name_classification`` leaf (C10 stage-2). They are imported
-# and re-exported above so the historical ``from .model import …`` path keeps
-# working.
-
-
-class Visibility(str, Enum):
-    PUBLIC = "public"  # default visibility / exported
-    HIDDEN = "hidden"  # __attribute__((visibility("hidden")))
-    ELF_ONLY = "elf_only"  # present in ELF symbol table, not in headers
-
-
-class ElfVisibility(str, Enum):
-    """ELF st_other visibility from .dynsym — separate from API-level Visibility."""
-
-    DEFAULT = "default"  # STV_DEFAULT
-    PROTECTED = "protected"  # STV_PROTECTED
-    HIDDEN = "hidden"  # STV_HIDDEN
-    INTERNAL = "internal"  # STV_INTERNAL
-
-
-class AccessLevel(str, Enum):
-    PUBLIC = "public"
-    PROTECTED = "protected"
-    PRIVATE = "private"
-
-
-class ParamKind(str, Enum):
-    VALUE = "value"
-    POINTER = "pointer"
-    REFERENCE = "reference"
-    RVALUE_REF = "rvalue_ref"
-
-
-class ScopeOrigin(str, Enum):
-    """Where a declaration's defining header sits relative to the
-    user-provided public-header set — the *Origin* axis of the two-axis
-    Linkage × Origin surface model (ADR-024 D1, ADR-015 schema v6).
-
-    Classification is opt-in: it is only meaningful when the caller
-    supplies a public-header set (``-H``/``--header``; ``scan`` also takes
-    ``--public-header-dir``).
-    Without one, every declaration is ``UNKNOWN`` and downstream behaviour
-    is unchanged.
-    """
-
-    PUBLIC_HEADER = "public_header"  # defined in a provided public header
-    PRIVATE_HEADER = "private_header"  # project header outside the public set
-    SYSTEM_HEADER = "system_header"  # toolchain/system header (/usr/include, ...)
-    GENERATED = "generated"  # machine-generated header (moc_*, *.pb.h, generated/ ...)
-    EXPORT_ONLY = "export_only"  # exported by the binary but absent from any header
-    UNKNOWN = "unknown"  # no public set, or no source location
-
-
-@dataclass
-class Param:
-    name: str
-    type: str
-    kind: ParamKind = ParamKind.VALUE
-    default: str | None = None  # has default value (value not preserved)
-    pointer_depth: int = 0  # nesting: T=0, T*=1, T**=2
-    is_restrict: bool = False  # restrict-qualified pointer parameter
-    is_va_list: bool = False  # parameter is va_list (variadic argument list)
-
-
-@dataclass
-class Function:
-    name: str  # demangled
-    mangled: str  # mangled symbol name
-    return_type: str
-    params: list[Param] = field(default_factory=list)
-    visibility: Visibility = Visibility.PUBLIC
-    is_virtual: bool = False
-    is_noexcept: bool = False
-    is_extern_c: bool = False
-    vtable_index: int | None = None
-    source_location: str | None = None  # "header.h:42"
-    is_static: bool = False
-    is_const: bool = False  # const qualifier on this
-    is_volatile: bool = False  # volatile qualifier on this
-    is_pure_virtual: bool = False
-    is_deleted: bool = False  # = delete; previously callable → BREAKING
-    deleted_from_dwarf: bool = False  # True when is_deleted was set via DW_AT_deleted
-    is_inline: bool = False  # inline keyword / attribute in header
-    access: AccessLevel = AccessLevel.PUBLIC  # public/protected/private
-    return_pointer_depth: int = 0  # T=0, T*=1, T**=2
-    elf_visibility: ElfVisibility | None = None  # ELF st_other (populated from .dynsym)
-    ref_qualifier: str = ""  # "" (none), "&" (lvalue), "&&" (rvalue)
-    # explicit specifier on constructors / conversion operators (DW_AT_explicit /
-    # castxml @explicit). Tri-state to keep "unknown" distinct from "implicit":
-    # - True  → source has `explicit` (or `explicit(true)`)
-    # - False → source does not have `explicit`
-    # - None  → snapshot loader does not know (older snapshots, dumpers that
-    #           don't capture this attribute). The diff must skip the
-    #           detector when either side is None to avoid false API_BREAK
-    #           findings from schema evolution.
-    is_explicit: bool | None = None
-    # Hidden-friend marker (in-class `friend` declaration, often inline).
-    # Tri-state to keep "unknown" distinct from "not a friend":
-    # - True  → declared as a friend inside some class body (castxml
-    #           ``befriending`` attribute on the class points to this fn).
-    # - False → not a friend declaration.
-    # - None  → dumper/loader could not determine (older snapshots, DWARF-
-    #           only path). Diff detectors skip when either side is None.
-    is_hidden_friend: bool | None = None
-    # Provenance (ADR-015, schema v6). source_header is the defining header
-    # (source_location with the line/col stripped); origin classifies it
-    # against the provided public-header set. Both are additive: missing on
-    # older snapshots and default to None / UNKNOWN.
-    source_header: str | None = None
-    origin: ScopeOrigin = ScopeOrigin.UNKNOWN
-    # C ellipsis (...) — variadic calls use a different convention on common
-    # ABIs (%al on SysV x86-64, stack args on Apple AArch64). Tri-state:
-    # None = dumper/loader does not know (older snapshots); diff skips then.
-    is_variadic: bool | None = None
-    # Semantic contract attributes (nonnull, noreturn, format, alloc_size,
-    # malloc, returns_nonnull, warn_unused_result, sentinel, ...), normalized
-    # spellings. None = not captured (older snapshots / dumpers without
-    # attribute support); [] = captured, none present. Diff skips on None.
-    contract_attributes: list[str] | None = None
-    # Dynamic exception specification spelling ("throw()", "throw(int)", ...).
-    # "" = captured, no dynamic spec; None = not captured. `noexcept` is NOT
-    # folded in here — it keeps its dedicated is_noexcept field and kinds.
-    exception_spec: str | None = None
-    # `[[deprecated]]`/`[[deprecated("msg")]]` (or castxml's `deprecation`
-    # attribute, which carries the same message text): a non-empty string is
-    # the message, "" is a bare `[[deprecated]]` with no message. Unlike most
-    # other tri-state fields here, None does NOT unambiguously mean
-    # "unsupported" — castxml also reports None for a genuinely
-    # non-deprecated declaration (there is no separate "deprecated with no
-    # info" state to distinguish it from). So the diff detector gates on
-    # header-tier confirmation at the *snapshot* level (mirroring
-    # Param.default/param_defaults's own header-tier-only gate), not by
-    # skipping a None on either side of a single pair — that would silently
-    # miss every real "gained/lost deprecated" transition, since one side of
-    # a real transition is always None (not-deprecated) by construction.
-    deprecated: str | None = None
-    # Explicit C++11 `override` specifier on a virtual method declaration.
-    # Tri-state like is_explicit/is_hidden_friend: True/False = captured;
-    # None = dumper/loader does not know (older snapshots, DWARF/symbols-only
-    # mode, or a non-virtual/non-method declaration for which the specifier
-    # is not applicable). Populated by both header backends since G31 Phase C
-    # (castxml's own `attributes` regex; clang's `OverrideAttr` child node).
-    is_override: bool | None = None
-    # Qualified name of the class whose body declares this friend (the
-    # `befriending` owner in castxml terms), e.g. "ns::Foo". None when the
-    # function is not a hidden friend, or the owner could not be resolved
-    # (older snapshots, DWARF-only path). Surface classification must key
-    # demotion off the *owner's* origin (system/private/public header), not
-    # just the friend function's own — a hidden friend can never produce an
-    # exported symbol by construction, but that is only a reason to skip the
-    # not-exported check, not a reason to skip the header-provenance check.
-    # Appended after all pre-existing fields (rather than inserted next to
-    # is_hidden_friend) so this additive field cannot shift the positional
-    # slot of any field that came before it (Codex review: Function is a
-    # public, non-keyword-only dataclass, so an insertion mid-list would
-    # silently rebind existing positional-constructor arguments instead of
-    # failing).
-    hidden_friend_owner: str | None = None
-    # ELF symbol *linkage* (st_info.bind — GLOBAL/WEAK/LOCAL/UNIQUE/OTHER),
-    # populated from .dynsym the same way elf_visibility is (see
-    # dumper_elf_symbols._populate_elf_visibility). None on a non-ELF
-    # platform, an older snapshot predating this field, or a declaration with
-    # no matching exported symbol (e.g. a public header-only inline that
-    # never made it into the dynamic symbol table). Distinguishes a WEAK
-    # COMDAT definition (e.g. an in-class-defined/`inline` member) from a
-    # GLOBAL/STRONG export on an otherwise-identical FUNC_REMOVED finding,
-    # which neither `visibility` nor `is_inline` alone can do — but this is
-    # PROVIDER-SIDE evidence only, about the *library's own build*, not a
-    # guarantee about consumers: a WEAK/COMDAT symbol does not by itself mean
-    # every consumer already carries its own copy (a public `extern template`
-    # declaration is the documented counterexample — the consumer TU
-    # deliberately does *not* instantiate, while the library's own explicit
-    # instantiation still emits a WEAK/COMDAT definition, so the consumer can
-    # hold an undefined reference to a symbol this field still reports as
-    # WEAK). See AGENTS.md's "Linkage-blind removal" entry for why a heavier
-    # removal-severity *demotion* keyed off this same fact was attempted and
-    # reverted for exactly this reason — this field only makes the fact
-    # visible on the model and matchable by a suppression selector, it does
-    # not itself change any verdict, and a suppression author must not treat
-    # WEAK alone as sufficient justification (see Suppression.binding's own
-    # docstring for the full caveat). Known gap, same as elf_visibility: a
-    # symbol-versioned bare name with mixed bindings across versions (e.g. a GLOBAL
-    # old-ABI @V1 and a WEAK default @@V2) collapses to whichever entry
-    # elf.symbol_map's last-write-wins dict happens to keep — see AGENTS.md's
-    # dedicated entry for this field.
-    elf_binding: SymbolBinding | None = None
-
-
-@dataclass
-class Variable:
-    name: str
-    mangled: str
-    type: str
-    visibility: Visibility = Visibility.PUBLIC
-    source_location: str | None = None
-    is_const: bool = False  # const-qualified type (write → SIGSEGV)
-    value: str | None = None  # initial value (compile-time constant, if known)
-    access: AccessLevel = AccessLevel.PUBLIC  # public/protected/private
-    elf_visibility: ElfVisibility | None = None  # ELF st_other (populated from .dynsym)
-    # Provenance (ADR-015, schema v6) — see Function.source_header.
-    source_header: str | None = None
-    origin: ScopeOrigin = ScopeOrigin.UNKNOWN
-    # Declared alignment in bits: an explicit alignas / __attribute__((aligned))
-    # override when present, else the variable's type's natural (computed)
-    # alignment when a dumper can resolve it. None = not captured (older
-    # snapshots / dumpers without support).
-    alignment_bits: int | None = None
-    # See Function.deprecated for the message-string convention.
-    deprecated: str | None = None
-    # See Function.elf_binding for the ELF-linkage rationale; same population
-    # path (dumper_elf_symbols._populate_elf_visibility).
-    elf_binding: SymbolBinding | None = None
-
-
-@dataclass
-class TypeField:
-    name: str
-    type: str
-    offset_bits: int | None = None
-    is_bitfield: bool = False
-    bitfield_bits: int | None = None
-    is_const: bool = False
-    is_volatile: bool = False
-    is_mutable: bool = False
-    access: AccessLevel = AccessLevel.PUBLIC
-    # Default member initializer expression, verbatim (value not evaluated).
-    # None = no initializer, or the dumper does not capture this (older
-    # snapshots / non-castxml producers). As with Function.deprecated, None
-    # is not unambiguously "unsupported" here — a real "gained/lost
-    # initializer" transition has one side genuinely None by construction —
-    # so the detector gates on header-tier confirmation at the *snapshot*
-    # level (mirroring Param.default/param_defaults) rather than skipping
-    # per-pair on either side being None.
-    default: str | None = None
-    # See Function.deprecated for the message-string convention.
-    deprecated: str | None = None
-
-
-@dataclass
-class RecordType:
-    """struct / class / union."""
-
-    name: str
-    kind: str  # "struct" | "class" | "union"
-    size_bits: int | None = None
-    alignment_bits: int | None = None
-    fields: list[TypeField] = field(default_factory=list)
-    bases: list[str] = field(default_factory=list)  # base class names
-    virtual_bases: list[str] = field(default_factory=list)
-    vtable: list[str] = field(default_factory=list)  # ordered vtable entries (mangled)
-    source_location: str | None = None
-    is_union: bool = False
-    is_opaque: bool = (
-        False  # incomplete type (forward-decl only; was complete → BREAKING)
-    )
-    # `final` class-key specifier. Tri-state to keep "unknown" distinct from
-    # "not final":
-    # - True  → declared `class C final { ... }` (castxml `final` attribute).
-    # - False → declared without `final`.
-    # - None  → dumper/loader could not determine (DWARF/symbols-only mode,
-    #           which carries no `final` information; older snapshots). The
-    #           diff skips the finality detector when either side is None to
-    #           avoid false findings from schema evolution / tier downgrade.
-    is_final: bool | None = None
-    # True when this RecordType is a class/struct template's own pattern body
-    # (e.g. the clang header backend's CXXRecordDecl nested inside a
-    # ClassTemplateDecl) rather than a concrete, instantiable type. Its field
-    # *names*/*types* are still real public surface, but it has no fixed
-    # layout for any one instantiation — detectors that need real
-    # size/offset data (e.g. DWARF layout backfill's name-based matching)
-    # must not treat it as an ordinary type. False for every non-clang
-    # producer (castxml/DWARF never emit an uninstantiated pattern this way).
-    is_template_pattern: bool = False
-    # True when *every* entry in `fields` was flattened up from an anonymous
-    # struct/union member by the clang header backend (clang emits an
-    # IndirectFieldDecl for each such member; see dumper_clang.py) -- not
-    # merely "at least one was" (Codex review): a mixed record like
-    # `struct Foo { union { int i; }; int tag; };` has an ordinary field
-    # (`tag`) with no such provenance guarantee, so the flag must be False
-    # for it too. DWARF's own record builder (dwarf_snapshot.py) now flattens
-    # *supported* anonymous aggregates too, but an unsupported producer/shape
-    # or a cached snapshot predating that flatten still legitimately leaves
-    # an all-anonymous record's DWARF view fieldless even though it carries
-    # the real size_bits — a structural signal the DWARF layout backfill
-    # needs to trust a
-    # bare-suffix (namespaced) match for this case without also trusting an
-    # ordinary record's coincidental match to an unrelated, fieldless type
-    # reached the same way. False for every non-clang producer (castxml
-    # computes real layout itself and is never backfilled; DWARF-only
-    # snapshots have no header view to flatten).
-    has_anonymous_aggregate_fields: bool = False
-    # Provenance (ADR-015, schema v6) — see Function.source_header.
-    source_header: str | None = None
-    origin: ScopeOrigin = ScopeOrigin.UNKNOWN
-    # ── Fine-grained layout descriptor (layout-closure work) ─────────────────
-    # All tri-state / optional so "unknown" (DWARF-only or symbols-only dumps,
-    # older snapshots) stays distinct from a real value; the layout detectors
-    # skip a comparison whenever either side is None/empty, avoiding false
-    # findings from schema evolution or an evidence-tier downgrade.
-    #
-    # Itanium "data size" (a.k.a. dsize/nvsize): the size occupied by the
-    # object's own members *excluding* trailing tail padding. A derived class
-    # may reuse a base's tail padding, so a change here can shift a derived
-    # layout even when ``size_bits`` (the padded sizeof) is unchanged.
-    data_size_bits: int | None = None
-    # C++ type traits that govern tail-padding reuse and how the type is passed
-    # by value (in registers vs. on the stack / via hidden reference).
-    is_standard_layout: bool | None = None
-    is_trivially_copyable: bool | None = None
-    # Bit offset of the vtable pointer within the object (0 for a simple
-    # polymorphic class; nonzero with virtual bases). None when the type is
-    # non-polymorphic or the dumper could not determine it. Introducing the
-    # first virtual function makes this go from None → 0 and shifts every field.
-    vptr_offset_bits: int | None = None
-    # Base-class subobject offsets: base name → bit offset within this object.
-    # Distinct from ``bases`` (declaration order only): a base can *move* (e.g.
-    # an empty-base-optimization is lost, or a member is inserted ahead of it)
-    # without the name list reordering. Empty when unknown.
-    base_offsets: dict[str, int] = field(default_factory=dict)
-    # Namespace/enclosing-class-qualified spelling (e.g. "mylib::detail::Impl"),
-    # set only when it differs from the bare ``name`` above. ``name`` itself
-    # stays bare (matching the DWARF backend, which has no cheaper way to
-    # qualify a struct name) so type-map lookups and DWARF/header merges keep
-    # matching by the same key across both backends; this field exists solely
-    # for namespace-aware checks (internal-leak detection, SYCL-queue param
-    # matching) that need to see the real namespace path. None when the type
-    # is at global scope or the dumper couldn't determine it (e.g. DWARF-only).
-    qualified_name: str | None = None
-    # Whether the class/struct declares at least one pure virtual function
-    # (making it abstract — cannot be instantiated). Tri-state like
-    # ``is_final``: True/False = captured (castxml's `abstract` attribute;
-    # clang's `definitionData.isAbstract` since G31 Phase C); None =
-    # dumper/loader could not determine (DWARF/symbols-only mode, older
-    # snapshots). The diff skips comparison when either side is None.
-    is_abstract: bool | None = None
-    # See Function.deprecated for the message-string convention.
-    deprecated: str | None = None
-
-
-@dataclass
-class EnumMember:
-    name: str
-    value: int
-
-
-@dataclass
-class EnumType:
-    name: str
-    members: list[EnumMember] = field(default_factory=list)
-    underlying_type: str = "int"
-    source_location: str | None = None
-    # Provenance (ADR-015, schema v6) — see Function.source_header.
-    source_header: str | None = None
-    origin: ScopeOrigin = ScopeOrigin.UNKNOWN
-    # `enum class` / `enum struct` (C++11 scoped enumeration) versus a plain
-    # C-style enum. Tri-state like RecordType.is_final: True/False = captured
-    # (castxml's `scoped` attribute); None = dumper/loader could not
-    # determine (DWARF/symbols-only mode, older snapshots, non-castxml
-    # header producers). The diff skips comparison when either side is None.
-    is_scoped: bool | None = None
-    # See Function.deprecated for the message-string convention.
-    deprecated: str | None = None
-    # Namespace/enclosing-class-qualified spelling, mirroring
-    # ``RecordType.qualified_name`` (same bare-``name``-collision motivation:
-    # PR #608 follow-up). ``name`` stays bare for the same DWARF-parity and
-    # type-map-key reasons documented on ``RecordType.qualified_name``. None
-    # when the enum is at global scope or the dumper couldn't determine it.
-    qualified_name: str | None = None
-
-
-@dataclass
-class DependencyInfo:
-    """Resolved transitive dependency graph and symbol bindings.
-
-    Populated when a snapshot is created with ``--follow-deps``.
-    """
-
-    nodes: list[dict[str, object]] = field(default_factory=list)
-    edges: list[dict[str, str]] = field(default_factory=list)
-    unresolved: list[dict[str, str]] = field(default_factory=list)
-    bindings_summary: dict[str, int] = field(default_factory=dict)
-    missing_symbols: list[dict[str, str]] = field(default_factory=list)
-
-
-@dataclass
-class ExtractionContract:
-    """ADR-050 D1 — profile/scope fingerprints proving two snapshots were
-    extracted under a comparable contract, plus the resolved per-field
-    inputs each fingerprint was computed from (so a mismatch report can show
-    *what* differs, not just that the hashes don't match).
-
-    Built by ``abicheck.comparability.compute_extraction_contract`` — never
-    constructed by hand outside tests. Both fingerprints are independently
-    optional: a symbols-only dump with no header-AST inputs but a real
-    a public-header set still attaches a
-    ``scope_fingerprint`` with ``profile_fingerprint=None`` (see that
-    module's docstring for the full rationale).
-    """
-
-    profile_fingerprint: str | None = None
-    scope_fingerprint: str | None = None
-    # Named resolved sub-inputs, one string per component, keyed the same way
-    # on both sides of a compare so a mismatch can be attributed to a specific
-    # field instead of an opaque hash. See ``comparability.PROFILE_FIELD_KEYS``
-    # / ``comparability.SCOPE_FIELD_KEYS`` for the recognized keys.
-    profile_fields: dict[str, str] = field(default_factory=dict)
-    scope_fields: dict[str, str] = field(default_factory=dict)
 
 
 @dataclass
@@ -1133,53 +668,29 @@ class AbiSnapshot:
         """
         if self._type_by_name is not None:
             return
-        func_map: dict[str, Function] = {}
-        dup_funcs: dict[str, int] = {}
-        for f in self.functions:
-            if f.mangled in func_map:
-                dup_funcs[f.mangled] = dup_funcs.get(f.mangled, 0) + 1
-            else:
-                func_map[f.mangled] = f
-        if dup_funcs:
-            _model_log.warning(
-                "Duplicate mangled symbols skipped (first-wins) in %s@%s: %s",
-                self.library,
-                self.version,
-                ", ".join(f"{k} (×{v + 1})" for k, v in dup_funcs.items()),
-            )
-        self._func_by_mangled = func_map
+        functions = build_first_wins_index(self.functions, lambda f: f.mangled)
+        self._warn_dropped("mangled symbols", functions.dropped)
+        self._func_by_mangled = functions.mapping
 
-        var_map: dict[str, Variable] = {}
-        dup_vars: dict[str, int] = {}
-        for v in self.variables:
-            if v.mangled in var_map:
-                dup_vars[v.mangled] = dup_vars.get(v.mangled, 0) + 1
-            else:
-                var_map[v.mangled] = v
-        if dup_vars:
-            _model_log.warning(
-                "Duplicate mangled variables skipped (first-wins) in %s@%s: %s",
-                self.library,
-                self.version,
-                ", ".join(f"{k} (×{v + 1})" for k, v in dup_vars.items()),
-            )
-        self._var_by_mangled = var_map
+        variables = build_first_wins_index(self.variables, lambda v: v.mangled)
+        self._warn_dropped("mangled variables", variables.dropped)
+        self._var_by_mangled = variables.mapping
 
-        type_map: dict[str, RecordType] = {}
-        dup_types: dict[str, int] = {}
-        for t in self.types:
-            if t.name in type_map:
-                dup_types[t.name] = dup_types.get(t.name, 0) + 1
-            else:
-                type_map[t.name] = t
-        if dup_types:
-            _model_log.warning(
-                "Duplicate type names skipped (first-wins) in %s@%s: %s",
-                self.library,
-                self.version,
-                ", ".join(f"{k} (×{v + 1})" for k, v in dup_types.items()),
-            )
-        self._type_by_name = type_map
+        types = build_first_wins_index(self.types, lambda t: t.name)
+        self._warn_dropped("type names", types.dropped)
+        self._type_by_name = types.mapping
+
+    def _warn_dropped(self, subject: str, dropped: dict[str, int]) -> None:
+        """Report the declarations a first-wins index had to drop, if any."""
+        if not dropped:
+            return
+        _model_log.warning(
+            "Duplicate %s skipped (first-wins) in %s@%s: %s",
+            subject,
+            self.library,
+            self.version,
+            describe_dropped(dropped),
+        )
 
     @property
     def function_map(self) -> dict[str, Function]:
